# api-gateway/internal/adapter/input/api/docs/swagger.yaml
definitions:
...
  order_request.CreateRequest:
    properties:
      address:
        type: string
      items:
        items:
          $ref: '#/definitions/order_request.ItemSchema'
        minItems: 1
        type: array
    required:
    - address
    - items
    type: object
  order_request.ItemSchema:
    properties:
      count:
        type: integer
      price:
        type: number
      product_id:
        type: string
    required:
    - count
    - price
    - product_id
    type: object
  order_response.DeliverySchema:
    properties:
      address:
        type: string
      arrived:
        type: string
      courier_id:
        type: string
    type: object
  order_response.ItemSchema:
    properties:
      count:
        type: integer
      price:
        type: number
      product_id:
        type: string
    type: object
  order_response.OrderResponse:
    properties:
      created:
        type: string
      customer_id:
        type: string
      delivery:
        $ref: '#/definitions/order_response.DeliverySchema'
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/order_response.ItemSchema'
        type: array
      status:
        type: string
      version:
        type: string
    type: object
  order_response.OrdersResponse:
    properties:
      orders:
        items:
          $ref: '#/definitions/order_response.OrderResponse'
        type: array
    type: object
  response.ErrorResponseDetail:
    properties:
      detail:
        type: string
    type: object
...
info:
  contact: {}
  description: This is the API Gateway for the Clean DDD application.
  title: Clean DDD App API Gateway
  version: "1.0"
paths:
...
  /couriers/me/orders:
    get:
      consumes:
      - application/json
      description: Get all current orders for the authenticated courier
      parameters:
      - in: query
        minimum: 1
        name: limit
        type: integer
      - in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of orders
          schema:
            $ref: '#/definitions/order_response.OrdersResponse'
        "401":
          description: Missing or invalid bearer token
          schema:
            $ref: '#/definitions/response.ErrorResponseDetail'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/response.ErrorResponseDetail'
      security:
      - CourierBearerAuth: []
      summary: Get courier orders
      tags:
      - couriers
...
  /orders:
    get:
      consumes:
      - application/json
      description: Get all orders for the authenticated customer
      parameters:
      - in: query
        minimum: 1
        name: limit
        type: integer
      - in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of orders
          schema:
            $ref: '#/definitions/order_response.OrdersResponse'
        "401":
          description: Missing or invalid bearer token
          schema:
            $ref: '#/definitions/response.ErrorResponseDetail'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/response.ErrorResponseDetail'
      security:
      - CustomerBearerAuth: []
      summary: Get customer orders
      tags:
      - orders
    post:
      consumes:
      - application/json
      description: Create a new order with items
      parameters:
      - description: Order details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/order_request.CreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: ' "Created with location header'
        "400":
          description: Invalid request format
          schema:
            $ref: '#/definitions/response.ErrorResponseDetail'
        "401":
          description: Missing or invalid bearer token
          schema:
            $ref: '#/definitions/response.ErrorResponseDetail'
        "422":
          description: Invalid item data
          schema:
            $ref: '#/definitions/response.ErrorResponseDetail'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/response.ErrorResponseDetail'
      security:
      - CustomerBearerAuth: []
      summary: Create a new order
      tags:
      - orders
  /orders/{id}/cancel:
    patch:
      consumes:
      - application/json
      description: Cancel an order by its ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: ' "No Content'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/response.ErrorResponseDetail'
        "401":
          description: Missing or invalid bearer token
          schema:
            $ref: '#/definitions/response.ErrorResponseDetail'
        "404":
          description: Order not found
          schema:
            $ref: '#/definitions/response.ErrorResponseDetail'
        "422":
          description: Invalid order ID format
          schema:
            $ref: '#/definitions/response.ErrorResponseDetail'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/response.ErrorResponseDetail'
      security:
      - CustomerBearerAuth: []
      summary: Cancel an order
      tags:
      - orders
  /orders/{id}/complete:
    patch:
      consumes:
      - application/json
      description: Mark an order as delivered (completed)
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ' "OK'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/response.ErrorResponseDetail'
        "401":
          description: Missing or invalid bearer token
          schema:
            $ref: '#/definitions/response.ErrorResponseDetail'
        "404":
          description: Order not found
          schema:
            $ref: '#/definitions/response.ErrorResponseDetail'
        "422":
          description: Invalid order ID format
          schema:
            $ref: '#/definitions/response.ErrorResponseDetail'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/response.ErrorResponseDetail'
      security:
      - CourierBearerAuth: []
      summary: Complete order
      tags:
      - orders
...
securityDefinitions:
  AdminAccessToken:
    description: Admin's access token.
    in: header
    name: X-Access-Token
    type: apiKey
  CourierBearerAuth:
    description: Courier's JWT token.
    in: header
    name: Authorization
    type: apiKey
  CustomerBearerAuth:
    description: Customer's JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"